+++ tempodb/encoding/vparquet/block_search.go
@@ -29,12 +29,28 @@ const (
 	StatusCodeUnset = "unset"
 	StatusCodeOK    = "ok"
 	StatusCodeError = "error"
+
+	KindUnspecified = "unspecified"
+	KindInternal    = "internal"
+	KindClient      = "client"
+	KindServer      = "server"
+	KindProducer    = "producer"
+	KindConsumer    = "consumer"
 )
 
 var StatusCodeMapping = map[string]int{
 	StatusCodeUnset: int(v1.Status_STATUS_CODE_UNSET),
 	StatusCodeOK:    int(v1.Status_STATUS_CODE_OK),
 	StatusCodeError: int(v1.Status_STATUS_CODE_ERROR),
+}
+
+var KindMapping = map[string]int{
+	KindUnspecified: int(v1.Span_SPAN_KIND_UNSPECIFIED),
+	KindInternal:    int(v1.Span_SPAN_KIND_INTERNAL),
+	KindClient:      int(v1.Span_SPAN_KIND_CLIENT),
+	KindServer:      int(v1.Span_SPAN_KIND_SERVER),
+	KindProducer:    int(v1.Span_SPAN_KIND_PRODUCER),
+	KindConsumer:    int(v1.Span_SPAN_KIND_CONSUMER),
 }
 
 // openForSearch consolidates all the logic for opening a parquet file
@@ -76,11 +92,6 @@ func (b *backendBlock) openForSearch(ctx context.Context, opts common.SearchOpti
 	span, _ := opentracing.StartSpanFromContext(ctx, "parquet.OpenFile")
 	defer span.Finish()
 	pf, err := parquet.OpenFile(readerAt, int64(b.meta.Size), o...)
-
-	if err == nil {
-		b.pf = pf
-		b.readerAt = backendReaderAt
-	}
 
 	return pf, backendReaderAt, err
 }