+++ tempodb/tempodb_search_test.go
@@ -19,6 +19,7 @@ import (
 	v1 "github.com/grafana/tempo/pkg/tempopb/trace/v1"
 	"github.com/grafana/tempo/pkg/traceql"
 	"github.com/grafana/tempo/pkg/util"
+	"github.com/grafana/tempo/pkg/util/math"
 	"github.com/grafana/tempo/pkg/util/test"
 	"github.com/grafana/tempo/tempodb/backend"
 	"github.com/grafana/tempo/tempodb/backend/local"
@@ -120,9 +121,10 @@ func testAdvancedTraceQLCompleteBlock(t *testing.T, blockVersion string) {
 
 		// collect some info about wantTr to use below
 		trueConditionsBySpan := [][]string{}
+		durationBySpan := []uint64{}
 		falseConditions := []string{
 			fmt.Sprintf("name=`%v`", test.RandomString()),
-			fmt.Sprintf("duration>%dh", rand.Intn(10)),
+			fmt.Sprintf("duration>%dh", rand.Intn(10)+1),
 			// status? can't really construct a status condition that's false for all spans
 		}
 		totalSpans := 0
@@ -136,13 +138,16 @@ func testAdvancedTraceQLCompleteBlock(t *testing.T, blockVersion string) {
 					trueC, falseC := conditionsForAttributes(s.Attributes, "span")
 
 					status := trace.StatusToString(s.Status.Code)
+					kind := trace.KindToString(s.Kind)
 					trueC = append(trueC, fmt.Sprintf("name=`%v`", s.Name))
 					trueC = append(trueC, fmt.Sprintf("duration=%dns", s.EndTimeUnixNano-s.StartTimeUnixNano))
 					trueC = append(trueC, fmt.Sprintf("status=%s", status))
+					trueC = append(trueC, fmt.Sprintf("kind=%s", kind))
+					trueC = append(trueC, trueResourceC...)
 
 					trueConditionsBySpan = append(trueConditionsBySpan, trueC)
-					trueConditionsBySpan = append(trueConditionsBySpan, trueResourceC)
 					falseConditions = append(falseConditions, falseC...)
+					durationBySpan = append(durationBySpan, s.EndTimeUnixNano-s.StartTimeUnixNano)
 				}
 			}
 		}
@@ -159,6 +164,8 @@ func testAdvancedTraceQLCompleteBlock(t *testing.T, blockVersion string) {
 			// spansets
 			{Query: fmt.Sprintf("{%s} && {%s}", rando(trueConditionsBySpan[0]), rando(trueConditionsBySpan[1]))},
 			{Query: fmt.Sprintf("{%s} || {%s}", rando(trueConditionsBySpan[0]), rando(falseConditions))},
+			{Query: fmt.Sprintf("{%s} && {%s} && {%s} && {%s} && {%s}", rando(trueConditionsBySpan[0]), rando(trueConditionsBySpan[0]), rando(trueConditionsBySpan[0]), rando(trueConditionsBySpan[0]), rando(trueConditionsBySpan[0]))},
+			{Query: fmt.Sprintf("{%s} || {%s} || {%s} || {%s} || {%s}", rando(falseConditions), rando(falseConditions), rando(falseConditions), rando(trueConditionsBySpan[0]), rando(falseConditions))},
 			{Query: fmt.Sprintf("{%s && %s} || {%s}", rando(falseConditions), rando(falseConditions), rando(trueConditionsBySpan[0]))},
 			// pipelines
 			{Query: fmt.Sprintf("{%s} | {%s}", rando(trueConditionsBySpan[0]), rando(trueConditionsBySpan[0]))},
@@ -169,10 +176,26 @@ func testAdvancedTraceQLCompleteBlock(t *testing.T, blockVersion string) {
 			// counts
 			{Query: fmt.Sprintf("{} | count() = %d", totalSpans)},
 			{Query: fmt.Sprintf("{} | count() != %d", totalSpans+1)},
-			{Query: fmt.Sprintf("{} | count() <= %d", totalSpans)},
-			{Query: fmt.Sprintf("{} | count() >= %d", totalSpans)},
-			// avgs
-			{Query: "{ } | avg(duration) > 0"}, // todo: make this better
+			{Query: fmt.Sprintf("{ true } && { true } | count() = %d", totalSpans)},
+			{Query: fmt.Sprintf("{ true } || { true } | count() = %d", totalSpans)},
+			{Query: fmt.Sprintf("{ %s && %s } | count() = 1", rando(trueConditionsBySpan[0]), rando(trueConditionsBySpan[0]))},
+			// avgs/min/max/sum
+			{Query: fmt.Sprintf("{ %s && %s } && { %s && %s } | avg(duration) = %dns",
+				rando(trueConditionsBySpan[0]), rando(trueConditionsBySpan[0]),
+				rando(trueConditionsBySpan[1]), rando(trueConditionsBySpan[1]),
+				(durationBySpan[0]+durationBySpan[1])/2)},
+			{Query: fmt.Sprintf("{ %s && %s } && { %s && %s } | min(duration) = %dns",
+				rando(trueConditionsBySpan[0]), rando(trueConditionsBySpan[0]),
+				rando(trueConditionsBySpan[1]), rando(trueConditionsBySpan[1]),
+				math.Min64(int64(durationBySpan[0]), int64(durationBySpan[1])))},
+			{Query: fmt.Sprintf("{ %s && %s } && { %s && %s } | max(duration) = %dns",
+				rando(trueConditionsBySpan[0]), rando(trueConditionsBySpan[0]),
+				rando(trueConditionsBySpan[1]), rando(trueConditionsBySpan[1]),
+				math.Max64(int64(durationBySpan[0]), int64(durationBySpan[1])))},
+			{Query: fmt.Sprintf("{ %s && %s } && { %s && %s } | sum(duration) = %dns",
+				rando(trueConditionsBySpan[0]), rando(trueConditionsBySpan[0]),
+				rando(trueConditionsBySpan[1]), rando(trueConditionsBySpan[1]),
+				durationBySpan[0]+durationBySpan[1])},
 		}
 		searchesThatDontMatch := []*tempopb.SearchRequest{
 			// conditions
@@ -198,6 +221,9 @@ func testAdvancedTraceQLCompleteBlock(t *testing.T, blockVersion string) {
 			// avgs
 			{Query: "{ } | avg(.dne) != 0"},
 			{Query: "{ } | avg(duration) < 0"},
+			{Query: "{ } | min(duration) < 0"},
+			{Query: "{ } | max(duration) < 0"},
+			{Query: "{ } | sum(duration) < 0"},
 		}
 
 		for _, req := range searchesThatMatch {
@@ -303,7 +329,7 @@ func runCompleteBlockSearchTest(t testing.TB, blockVersion string, runner runner
 	}, log.NewNopLogger())
 	require.NoError(t, err)
 
-	c.EnableCompaction(&CompactorConfig{
+	c.EnableCompaction(context.Background(), &CompactorConfig{
 		ChunkSizeBytes:          10,
 		MaxCompactionRange:      time.Hour,
 		BlockRetention:          0,
@@ -433,7 +459,6 @@ func searchTestSuite() (
 	searchesThatMatch []*tempopb.SearchRequest,
 	searchesThatDontMatch []*tempopb.SearchRequest,
 ) {
-
 	id = test.ValidTraceID(nil)
 
 	start = 1000
@@ -493,7 +518,7 @@ func searchTestSuite() (
 								TraceId:           id,
 								Name:              "RootSpan",
 								StartTimeUnixNano: uint64(1000 * time.Second),
-								EndTimeUnixNano:   uint64(1001 * time.Second),
+								EndTimeUnixNano:   uint64(1002 * time.Second),
 								Status:            &v1.Status{},
 							},
 						},
@@ -506,7 +531,7 @@ func searchTestSuite() (
 	expected = &tempopb.TraceSearchMetadata{
 		TraceID:           util.TraceIDToHexString(id),
 		StartTimeUnixNano: uint64(1000 * time.Second),
-		DurationMs:        1000,
+		DurationMs:        2000,
 		RootServiceName:   "RootService",
 		RootTraceName:     "RootSpan",
 	}
@@ -518,7 +543,7 @@ func searchTestSuite() (
 		},
 		{
 			MinDurationMs: 999,
-			MaxDurationMs: 1001,
+			MaxDurationMs: 2001,
 		},
 		{
 			Start: 1000,
@@ -573,7 +598,7 @@ func searchTestSuite() (
 	// Excludes
 	searchesThatDontMatch = []*tempopb.SearchRequest{
 		{
-			MinDurationMs: 1001,
+			MinDurationMs: 2001,
 		},
 		{
 			MaxDurationMs: 999,